Fix testsupport
