+++ Sources/PointFree/Environment.swift
@@ -18,10 +18,10 @@ public struct Environment {
   public var envVars = EnvVars()
   public var episodes = { [Episode]() }
   public var features = [Feature].allFeatures
-  public var gitHub = GitHub.Client(clientId: "", clientSecret: "", logger: nil)
+  public var gitHub: GitHub.Client!
   public var logger = Logger()
   public var mailgun = Mailgun.live
   public var renderHtml: ([Node]) -> String = Html.render
-  public var stripe = Stripe.Client(logger: nil, secretKey: "")
+  public var stripe: Stripe.Client!
   public var uuid: () -> UUID = UUID.init
 }
